jesd204: don't register devices with `jesd204-device` DT prop
Many drivers will want to function outside of the JESD204 framework
context. In that case, registering with the JESD204 framework may cause
problems, because [pretty likely] there is no device-tree entry for the
device, in which case registering with the framework would cause an error.

This change implements a mechanism to make registering with the framework
to be a "pass-through" mechanism, and the driver will function without the
framework ever knowing about it.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>