jesd204: collect all devices on framework init
The overall mode of operation for the framework will be that upon
initialization, the framework will collect all data (from device-trees)
for all JESD204 devices, and build internal representations for all of
them, and their respective links/topologies.

That way, when the kernel probes these devices, the JESD204 framework can
update states for these devices, links &topologies and can catch various
errors/omissions that drivers may have.
This makes it easier to see which device in a topology (of JESD204
devices) has updated/initialized and which error-ed.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>