drivers: jesd204-core: keep track of top JESD204 topologies
This change implements a differentiation between normal/node JESD204
devices and top JESD204 devices.

In a JESD204 topology, at least one device must be a top-level. This
device is the device for which all the JESD204 chain/topology must
synchronize (via this JESD204 framework) so that it can start/operate.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>