arch: arm64: add reference zynqmp-zcu102-rev10-adin1300 example
This change adds reference device-tree `zynqmp-zcu102-rev10-adin1300` which
extends the Xilinx ZCU102 base device-tree to support the ADIN1300.

On the GEM0 path [which is connected to HPC0 on the board], there's an
RGMII interface to the board, and a GMII2RGMII FPGA converter.

On the GEM1 path [connected to HPC1 on the board], there's an MII
interface.

At this time, GEM0 works out-of-the-box as the ADIN1300 PHY chip is in the
default state of RGMII. More stuff will be needed to extend the
functionality of the driver to switch to other MII modes.

This device-tree change will work with just the Generic PHY driver in the
kernel.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit d959f09a3eef6cfcfc9b87ab5f06ed55ebd4f890)