net: phy: adin: support PHY mode converters
Sometimes, the connection between a MAC and PHY is done via a mode
converter. An example is a GMII-to-RGMII converter, which would mean that
the MAC operates in GMII mode while the PHY operates in RGMII. In this
case there is a discrepancy between what the MAC expects & what the PHY
expects and both need to be configured in their respective modes.

Sometimes, this converter is specified in the device-tree. But, other times
it can be left unspecified.

This patch also adds support for a `phy-mode-internal` device-tree property
that can be used in these (implicit convert) cases. The internal PHY mode
will be used to specify the correct register settings for the PHY.

`fwnode_handle` is used, since this property may be specified via ACPI as
well in other setups, but testing has been done in DT context.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit ad8d8135d7c4c7450fbacb1c0863abfd33d6a35d)