net: phy: adin: add compat layer
This change adds a compatibility layer to make sure that this works on
older (than 5.0) Linux versions.

Tested that it builds down to 4.1

We don't have a way to test 4.1 as that requires a Xilinx base-system from
that era.
Tested that this builds on 4.1, so hopefully the compat layer works. There
were many changes around the 4.4 version.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>
(cherry picked from commit 5423d4cc11736180faa89a56b0eefb334970e3bb)