iio: adc: adrv9001: WIP - Eliminate fileIO usage
A generic device driver must not assume a libc environment. Especially not
the presence of fileIO. This work in progress patch currently only
removes the code for Linux.
In general there is a lot of boilerplate code between these functions,
which is sub-optimal. The fileIO part which loads the file into a
buffer and retrieves the size must be split into a platform function.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>