iio: adc: ad9680: register enable & disable link callbacks
This change moves the init & uninit of the JESD204 links to the framework.
The move is not perfect, in the sense that it works, but not completely
tested that no regression is introduced.

The aim is to move parts of the driver into the framework into a manner
that is reviewable.

The `ad9680_jesd204_links` is copied from the `ad9680_setup()` as a static
configuration. The only real need for it (at this point in time), is to
pass `ARRAY_SIZE(ad9680_jesd204_links)` to `num_links`, so that the
framework knows that this device has 1 JESD204 link it needs to operate.

The framework will call `ad9680_jesd204_enable_link()` after the device
registers with the JESD204 framework.

The way `ad9680_jesd204_links` is (right now) is pretty bad, as it
completely ignores device-tree configuration; and relies on the fact that
the DT has a specific configuration.
At a later point in time, this should be made configurable again.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>