iio: adc: ad9680: split enable/disable clocks from ad9680_request_clks()
The disable & enable clocks must be their own states in the JESD204
framework, so that (when we bring down & up) a JESD204 link, we can go
through a "disable/enable all clocks for a given link".

Disabling/enabling all clocks for JESD204 link seems to be important.
This change splits these ops into their own callbacks. Note that some
clocks may be important to be in the enable/disable link phase (like the
`conv->lane_clk`).

To keep backwards compatibility with the non-framework behavior, the
clock enables are also copied in the `ad9680_setup()` function after the
quirky goto statement.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>