iio: adc: ad9680: move enable/disable clocks to JESD204 framework
The disable & enable clocks must be their own states, so that (when we
bring down & up) a JESD204 link, we can go through a "disable/enable all
clocks for a given link".

Disabling/enabling all clocks for JESD204 link seems to be important.
And this change split these into their own callbacks. Note that some
clocks may be important to be in the enable/disable link phase (like the
`conv->lane_clk`).
Even though, in this case, the `conv->lane_clk` is essentially
enabling/disabling the link on the `axi_jesd204_rx` peripheral.
This may change later.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>