iio: fingerprint: adux1060: Add basic support for ADUX1060
The ADUX1060 is a controller for use in low power ultrasound fingerprint
imaging applications. The ADUX1060 integrates a flexible ultrasound
imaging Analog Front End (AFE) and ARM Cortex M3 microcontroller digital
core operating at up to 120MHz. Connection to the host processor is via
a 30MHz SPI interface.

This basic patch resets the ADUX1060 and writes the bootloader via the
SPI interface by following the strict boot protocol described in the
Bootloader section of the datasheet.

The SPI protocol must include a delay of 250us between the falling edge
of the CS signal and the first edge of the SCLK signal, this allows the
ADUX1060 to wake from hibernate and prepare to receive the SPI command.
For this reason, a different GPIO had to be used as chip select.

After a successful boot, the registers can be read/written by using the
debugfs_reg_access.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>