iio: fingerprint: adux1060: Add support for untouch and touch scans
Before an image is being captured, an untouch and a touch scan needs to
be performed.

Before initializing an untouch scan, all interrupts are disabled except
the untouch irq (UNTOU_IRQ_EN), afterwards the driver waits for one
interrupt to occur.

Before initalizing a touch scan, only the touch interrupt is enabled
(TOU_IRQ_EN) and the driver needs to wait for one or several interrupts.
The number of interrupts that will occur is equal with the total number
of rows (by default 120) divided with the number of image rows
(NUM_IMG_ROWS from SCAN_CFG[31:24] reg). By default, the number of image
rows is set to 60, so the driver will wait for two interrupts to occur.

Signed-off-by: Stefan Popa <stefan.popa@analog.com>