iio: fingerprint: adux1060: convert cs delay to dummy transfer
Initially I thought that a change is required to the SPI framework, but
after looking a bit inside it, it looks like we can introduce transfers
with no data (0 length & NULL buffers) and just a `delay`, which seems to
be perfect for the case of waking the device from sleep mode.

If we add a transfer with no data but 1 delay, the loop will just sleep 1
transfer (which is a dummy) and all will be fine from their on.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>