iio: axi-adc: add support for LVDS delay calibration
The lanes between the AXI ADC IP core and the ASIC (client) ADC require
delays, in order to operate smoothly.

This change adds a basic calibration procedure that is done on the FPGA
side (inside the AXI ADC IP core) that tries to find the best delay to be
set for all lanes.
Later, this could be extended to try to determine delays per-lane, as that
could provide more flexibility when designing board layouts.

Signed-off-by: Michael Hennerich <michael.hennerich@analog.com>
Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>
Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>