hwmon: (ltc2945): add support for sense resistor
The sense resistor is a parameter of the board. It should be configured in
the driver via a device-tree / ACPI property, so that the proper current
measurements can be done in the driver.

It shouldn't be necessary that userspace need to know about the value of
the resistor. It makes things a bit harder to make the application code
portable from one board to another.

This change implements support for the sense resistor to be configured from
DT/ACPI and used in current calculations.
Also, the maximum power and current that can be represented by the driver
are scaled with the value of the sense resistor.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>