gpio: logicvc: remove unneeded platform_set_drvdata() call.
In the gpio-logicvc driver found in drivers/gpio, the call to platform_set_drvdata was removed. This function call was found to be unnecessary as the associated structure, 'logicvc_gpio', is defined within the same file and doesn't utilize a getter function. Therefore, it only sets values and does not retrieve them. This change simplifies the code without altering functionality.

Reviewed-by: Alexandru Ardelean <alex@shruggie.ro>
Signed-off-by: Andrei Coardos <traiandrei@yahoo.com>