hwmon: emc1403: use s32 type for emc1403_conv_time table
This is part of a rework for the find_closest() macro to become a function.
The intent is to make all arrays (used for find_closest() to be signed
32-bit.

This change renames the global variable from 'ina3221_conv_time' to
'emc1403_conv_time'. This looks like a copy+paste left-over.

And it converts the type of the elements in the array from u16 to s32.

Signed-off-by: Alexandru Ardelean <alex@shruggie.ro>